requirements.txt
Flask-Cors==3.0.10
  - Flask [required: >=0.9, installed: 2.1.2]
  - Six [required: Any, installed: 1.16.0]
google-search-results==2.4.2
  - requests [required: Any, installed: 2.28.1]
grpc-gateway-protoc-gen-openapiv2==0.1.0
  - googleapis-common-protos [required: Any, installed: 1.56.1]
langchain==0.0.112
  - aiohttp [required: >=3.8.3,<4.0.0, installed: 3.8.4]
    - aiosignal [required: >=1.1.2, installed: 1.3.1]
      - frozenlist [required: >=1.1.0, installed: 1.3.3]
    - async-timeout [required: >=4.0.0a3,<5.0, installed: 4.0.2]
    - attrs [required: >=17.3.0, installed: 21.4.0]
    - charset-normalizer [required: >=2.0,<4.0, installed: 2.1.0]
    - frozenlist [required: >=1.1.1, installed: 1.3.3]
    - multidict [required: >=4.5,<7.0, installed: 6.0.4]
    - yarl [required: >=1.0,<2.0, installed: 1.8.2]
      - idna [required: >=2.0, installed: 3.3]
      - multidict [required: >=4.0, installed: 6.0.4]
  - dataclasses-json [required: >=0.5.7,<0.6.0, installed: 0.5.7]
    - marshmallow [required: >=3.3.0,<4.0.0, installed: 3.19.0]
      - packaging [required: >=17.0, installed: 21.3]
    - marshmallow-enum [required: >=1.5.1,<2.0.0, installed: 1.5.1]
      - marshmallow [required: >=2.0.0, installed: 3.19.0]
        - packaging [required: >=17.0, installed: 21.3]
    - typing-inspect [required: >=0.4.0, installed: 0.8.0]
      - mypy-extensions [required: >=0.3.0, installed: 0.4.3]
      - typing-extensions [required: >=3.7.4, installed: 3.10.0.2]
  - numpy [required: >=1,<2, installed: 1.21.6]
  - pydantic [required: >=1,<2, installed: 1.9.0]
  - PyYAML [required: >=6,<7, installed: 6.0]
  - requests [required: >=2,<3, installed: 2.28.1]
  - SQLAlchemy [required: >=1,<2, installed: 1.4.36]
  - tenacity [required: >=8.1.0,<9.0.0, installed: 8.2.2]
lz4==4.3.2
openai==0.27.2
  - aiohttp [required: Any, installed: 3.8.4]
    - aiosignal [required: >=1.1.2, installed: 1.3.1]
      - frozenlist [required: >=1.1.0, installed: 1.3.3]
    - async-timeout [required: >=4.0.0a3,<5.0, installed: 4.0.2]
  - requests [required: >=2.20, installed: 2.28.1]
    - attrs [required: >=17.3.0, installed: 21.4.0]
    - charset-normalizer [required: >=2.0,<4.0, installed: 2.1.0]
    - frozenlist [required: >=1.1.1, installed: 1.3.3]
    - multidict [required: >=4.5,<7.0, installed: 6.0.4]
    - yarl [required: >=1.0,<2.0, installed: 1.8.2]
      - idna [required: >=2.0, installed: 3.3]
      - multidict [required: >=4.0, installed: 6.0.4]
  - requests [required: >=2.20, installed: 2.28.1]
  - tqdm [required: Any, installed: 4.65.0]
pinecone-client==2.2.1
  - dnspython [required: >=2.0.0, installed: 2.3.0]
  - loguru [required: >=0.5.0, installed: 0.6.0]
  - numpy [required: Any, installed: 1.21.6]
  - python-dateutil [required: >=2.5.3, installed: 2.8.2]
  - pyyaml [required: >=5.4, installed: 6.0]
  - requests [required: >=2.19.0, installed: 2.28.1]
  - tqdm [required: >=4.64.1, installed: 4.65.0]
  - typing-extensions [required: >=3.7.4, installed: 3.10.0.2]
  - urllib3 [required: >=1.21.1, installed: 1.26.9]